Separate component totalizer from table
