+++ react/components/Totalizers/TotalizerIcon.js
