+++ react/components/Totalizers/TotalizerLabel.js
