+++ react/components/Totalizers/TotalizerValue.js
